Revert dependencies for connection.h and impl/connection.h
